# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kasir Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  createdAt: string;
}

export interface CategoryInput {
  name: string;
}

export interface Product {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  stock?: number | null;
  isActive: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  stock?: number | null;
  isActive?: boolean;
}

export interface ProductUpdate {
  name?: string;
  price?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  stock?: number | null;
  isActive?: boolean;
}

export type CustomerMembershipType = typeof CustomerMembershipType[keyof typeof CustomerMembershipType];


export const CustomerMembershipType = {
  member: 'member',
  non_member: 'non_member',
} as const;

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  membershipType: CustomerMembershipType;
  points: number;
  totalSpent?: number;
  createdAt: string;
}

export type CustomerInputMembershipType = typeof CustomerInputMembershipType[keyof typeof CustomerInputMembershipType];


export const CustomerInputMembershipType = {
  member: 'member',
  non_member: 'non_member',
} as const;

export interface CustomerInput {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  membershipType: CustomerInputMembershipType;
}

export type CustomerUpdateMembershipType = typeof CustomerUpdateMembershipType[keyof typeof CustomerUpdateMembershipType];


export const CustomerUpdateMembershipType = {
  member: 'member',
  non_member: 'non_member',
} as const;

export interface CustomerUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  membershipType?: CustomerUpdateMembershipType;
  points?: number;
}

export type TransactionPaymentMethod = typeof TransactionPaymentMethod[keyof typeof TransactionPaymentMethod];


export const TransactionPaymentMethod = {
  cash: 'cash',
  debit_card: 'debit_card',
  credit_card: 'credit_card',
  qris: 'qris',
  transfer: 'transfer',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  completed: 'completed',
  voided: 'voided',
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  cashierName: string;
  paymentMethod: TransactionPaymentMethod;
  subtotal: number;
  discount?: number;
  tax: number;
  total: number;
  amountPaid?: number;
  change?: number;
  pointsEarned: number;
  pointsUsed?: number;
  status: TransactionStatus;
  createdAt: string;
}

export interface TransactionItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  subtotal: number;
}

export type TransactionDetail = Transaction & {
  items: TransactionItem[];
};

export type TransactionInputPaymentMethod = typeof TransactionInputPaymentMethod[keyof typeof TransactionInputPaymentMethod];


export const TransactionInputPaymentMethod = {
  cash: 'cash',
  debit_card: 'debit_card',
  credit_card: 'credit_card',
  qris: 'qris',
  transfer: 'transfer',
} as const;

export interface TransactionItemInput {
  productId: number;
  quantity: number;
  price: number;
}

export interface TransactionInput {
  /** @nullable */
  customerId?: number | null;
  cashierName: string;
  paymentMethod: TransactionInputPaymentMethod;
  discount?: number;
  amountPaid?: number;
  pointsUsed?: number;
  items: TransactionItemInput[];
}

export interface DashboardStats {
  totalRevenueToday: number;
  totalRevenueMonth: number;
  transactionsToday: number;
  transactionsMonth: number;
  totalCustomers: number;
  totalProducts: number;
  newCustomersThisMonth: number;
  averageTransactionValue?: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  /** @nullable */
  imageUrl?: string | null;
  totalSold: number;
  totalRevenue: number;
}

export interface RevenueChartPoint {
  date: string;
  revenue: number;
  transactions: number;
}

export type ListProductsParams = {
search?: string;
categoryId?: number;
isActive?: boolean;
};

export type ListCustomersParams = {
search?: string;
membershipType?: ListCustomersMembershipType;
};

export type ListCustomersMembershipType = typeof ListCustomersMembershipType[keyof typeof ListCustomersMembershipType];


export const ListCustomersMembershipType = {
  member: 'member',
  non_member: 'non_member',
} as const;

export type LookupCustomerParams = {
phone: string;
};

export type ListTransactionsParams = {
startDate?: string;
endDate?: string;
paymentMethod?: string;
limit?: number;
offset?: number;
};

export type GetTopProductsParams = {
limit?: number;
};

export type GetRecentTransactionsParams = {
limit?: number;
};

export type GetRevenueChartParams = {
days?: number;
};