# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kasir Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Products ──────────────────────────────────────────────
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: membershipType
          in: query
          schema:
            type: string
            enum: [member, non_member]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/lookup:
    get:
      operationId: lookupCustomer
      tags: [customers]
      summary: Lookup customer by phone number (for cashier)
      parameters:
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found

  # ─── Transactions ──────────────────────────────────────────
...
  /dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
  /dashboard/recent-transactions:
    get:
      operationId: getRecentTransactions
      tags: [dashboard]
      summary: Get recent transactions for dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
  /dashboard/revenue-chart:
    get:
      operationId: getRevenueChart
      tags: [dashboard]
      summary: Get daily revenue for the last N days
      parameters:
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Revenue chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueChartPoint"
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string

    # ─── Product ───────────────────────────────────────────
...
    Customer:
      type: object
      required: [id, name, membershipType, points, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        membershipType:
          type: string
          enum: [member, non_member]
        points:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, membershipType]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        membershipType:
          type: string
          enum: [member, non_member]
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        membershipType:
          type: string
          enum: [member, non_member]
        points:
          type: integer

    # ─── Transaction ───────────────────────────────────────
    Transaction:
      type: object
      required: [id, cashierName, paymentMethod, subtotal, tax, total, pointsEarned, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        cashierName:
          type: string
        paymentMethod:
          type: string
          enum: [cash, debit_card, credit_card, qris, transfer]
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        pointsEarned:
          type: integer
        pointsUsed:
          type: integer
        status:
          type: string
          enum: [completed, voided]
        createdAt:
          type: string
...
    TopProduct:
      type: object
      required: [productId, productName, totalSold, totalRevenue]
      properties:
        productId:
          type: integer
        productName:
          type: string
        imageUrl:
          type: ["string", "null"]
        totalSold:
          type: integer
        totalRevenue:
          type: number
    RevenueChartPoint:
      type: object
      required: [date, revenue, transactions]
      properties:
        date:
          type: string
        revenue:
          type: number
        transactions:
          type: integer