# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kasir Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerMembershipType } from './customerMembershipType';

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  membershipType: CustomerMembershipType;
  points: number;
  totalSpent?: number;
  createdAt: string;
}