# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Kasir Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateMembershipType } from './customerUpdateMembershipType';

export interface CustomerUpdate {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  membershipType?: CustomerUpdateMembershipType;
  points?: number;
}